pyproject.toml
[tool.poetry]
name = "blog"
version = "0.1.0"
description = ""
authors = ["Evan <evan.vilagov@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
gunicorn = "^20.1.0"
python-dotenv = "^0.17.1"
Django = "^3.2.4"
psycopg2-binary = "^2.9.1"
djangorestframework = "^3.12.4"

[tool.poetry.dev-dependencies]
flake8 = "^3.9.1"
pylint = "^2.8.3"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
